Move createRejectedPromise to internal
