+++ src/react/internal/cache/QueryReference.ts
@@ -10,10 +10,12 @@ import type {
   WatchQueryOptions,
 } from "@apollo/client";
 import type { MaybeMasked } from "@apollo/client/masking";
-import { createRejectedPromise } from "@apollo/client/utilities";
 import { wrapPromiseWithState } from "@apollo/client/utilities";
 import type { DecoratedPromise } from "@apollo/client/utilities/internal";
-import { createFulfilledPromise } from "@apollo/client/utilities/internal";
+import {
+  createFulfilledPromise,
+  createRejectedPromise,
+} from "@apollo/client/utilities/internal";
 import { invariant } from "@apollo/client/utilities/invariant";
 
 import type { QueryKey } from "./types.js";
