+++ src/utilities/internal/index.ts
@@ -21,6 +21,7 @@ export { cloneDeep } from "./cloneDeep.js";
 export { compact } from "./compact.js";
 export { createFragmentMap } from "./createFragmentMap.js";
 export { createFulfilledPromise } from "./createFulfilledPromise.js";
+export { createRejectedPromise } from "./createRejectedPromise.js";
 export { DeepMerger } from "./DeepMerger.js";
 export { getDefaultValues } from "./getDefaultValues.js";
 export { getFragmentFromSelection } from "./getFragmentFromSelection.js";
