+++ src/utilities/promises/decoration.ts
@@ -5,18 +5,6 @@ import type {
   RejectedPromise,
 } from "@apollo/client/utilities/internal";
 
-export function createRejectedPromise<TValue = unknown>(reason: unknown) {
-  const promise = Promise.reject(reason) as RejectedPromise<TValue>;
-
-  // prevent potential edge cases leaking unhandled error rejections
-  promise.catch(() => {});
-
-  promise.status = "rejected";
-  promise.reason = reason;
-
-  return promise;
-}
-
 export function isStatefulPromise<TValue>(
   promise: Promise<TValue>
 ): promise is DecoratedPromise<TValue> {
